+++ src-tauri/src/lib.rs
@@ -44,6 +44,7 @@ pub fn run() {
         .plugin(tauri_plugin_store::Builder::default().build())
         .plugin(tauri_plugin_updater::Builder::new().build())
         .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_global_shortcut::Builder::new().build())
         .manage(discord::DiscordState::default())
         .invoke_handler(tauri::generate_handler![
             tmdb::tmdb_search,